Add files via upload
EMPLOYEE DATA ANALYSIS 
This is an analysis on an employee data for a company with frequent attrition of employees. We would be analyzing the data to check the reason for the attrition and predict whether there would be more employees leaving the company. We would go ahead to import libraries needed for the analysis, upload and read the data and continue with the Exploratory Data Analysis to help unravel the reason for continuous employee attrition.